pyproject.toml
[project]
name = "object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "ultralytics>=8.3.233",
]
